+++ tests/local/test_ScraperExtractFromTextTest.py
@@ -1372,6 +1372,71 @@ class ScraperExtractFromText(unittest.TestCase):
                 },
             ),
         ],
+        "juriscraper.opinions.united_states.state.idaho_civil": [
+            (
+                "\n           ACADEMY ASSOCIATION, INC.,          )\na North Carolina corporation; and   )\nSPOKANE TEACHERS CREDIT             )\nUNION, a Washington credit union,   )\n                                    )\n       Defendants.                  )\n____________________________________)\n\n\n      Appeal from the District Court of the First Judicial District of the State of Idaho,\n      Bonner County. Lamont C. Berecz, District Judge.\n\n      The district court’s judgment is affirmed.\n\n      Condon Tobin Sladek Thornton Nerenberg, PLLC, Dallas, TX and Parsons Behle &\n      Latimer, Boise, attorneys for Appellant, Global Growth Holdings, Inc. Jared T.S.\n      Pace submitted argument on the briefs.\n\n      Givens Pursley, LLP, Boise, attorneys for Respondent, Erie Properties, LLC.\n      Morgan D. Goodin submitted argument on the briefs.\n",
+                {
+                    "Docket": {
+                        "appeal_from_str": "District Court of the First Judicial District of the State of Idaho, Bonner County",
+                    },
+                    "OriginatingCourtInformation": {
+                        "assigned_to_str": "Lamont C. Berecz"
+                    },
+                },
+            ),
+        ],
+        "juriscraper.opinions.united_states.state.idaho_criminal": [
+            (
+                "\n           ACADEMY ASSOCIATION, INC.,          )\na North Carolina corporation; and   )\nSPOKANE TEACHERS CREDIT             )\nUNION, a Washington credit union,   )\n                                    )\n       Defendants.                  )\n____________________________________)\n\n\n      Appeal from the District Court of the First Judicial District of the State of Idaho,\n      Bonner County. Lamont C. Berecz, District Judge.\n\n      The district court’s judgment is affirmed.\n\n      Condon Tobin Sladek Thornton Nerenberg, PLLC, Dallas, TX and Parsons Behle &\n      Latimer, Boise, attorneys for Appellant, Global Growth Holdings, Inc. Jared T.S.\n      Pace submitted argument on the briefs.\n\n      Givens Pursley, LLP, Boise, attorneys for Respondent, Erie Properties, LLC.\n      Morgan D. Goodin submitted argument on the briefs.\n",
+                {
+                    "Docket": {
+                        "appeal_from_str": "District Court of the First Judicial District of the State of Idaho, Bonner County",
+                    },
+                    "OriginatingCourtInformation": {
+                        "assigned_to_str": "Lamont C. Berecz"
+                    },
+                },
+            ),
+        ],
+        "juriscraper.opinions.united_states.state.idahoctapp_civil": [
+            (
+                "\n           ACADEMY ASSOCIATION, INC.,          )\na North Carolina corporation; and   )\nSPOKANE TEACHERS CREDIT             )\nUNION, a Washington credit union,   )\n                                    )\n       Defendants.                  )\n____________________________________)\n\n\n      Appeal from the District Court of the First Judicial District of the State of Idaho,\n      Bonner County. Lamont C. Berecz, District Judge.\n\n      The district court’s judgment is affirmed.\n\n      Condon Tobin Sladek Thornton Nerenberg, PLLC, Dallas, TX and Parsons Behle &\n      Latimer, Boise, attorneys for Appellant, Global Growth Holdings, Inc. Jared T.S.\n      Pace submitted argument on the briefs.\n\n      Givens Pursley, LLP, Boise, attorneys for Respondent, Erie Properties, LLC.\n      Morgan D. Goodin submitted argument on the briefs.\n",
+                {
+                    "Docket": {
+                        "appeal_from_str": "District Court of the First Judicial District of the State of Idaho, Bonner County",
+                    },
+                    "OriginatingCourtInformation": {
+                        "assigned_to_str": "Lamont C. Berecz"
+                    },
+                },
+            ),
+        ],
+        "juriscraper.opinions.united_states.state.idahoctapp_criminal": [
+            (
+                "\n           ACADEMY ASSOCIATION, INC.,          )\na North Carolina corporation; and   )\nSPOKANE TEACHERS CREDIT             )\nUNION, a Washington credit union,   )\n                                    )\n       Defendants.                  )\n____________________________________)\n\n\n      Appeal from the District Court of the First Judicial District of the State of Idaho,\n      Bonner County. Lamont C. Berecz, District Judge.\n\n      The district court’s judgment is affirmed.\n\n      Condon Tobin Sladek Thornton Nerenberg, PLLC, Dallas, TX and Parsons Behle &\n      Latimer, Boise, attorneys for Appellant, Global Growth Holdings, Inc. Jared T.S.\n      Pace submitted argument on the briefs.\n\n      Givens Pursley, LLP, Boise, attorneys for Respondent, Erie Properties, LLC.\n      Morgan D. Goodin submitted argument on the briefs.\n",
+                {
+                    "Docket": {
+                        "appeal_from_str": "District Court of the First Judicial District of the State of Idaho, Bonner County",
+                    },
+                    "OriginatingCourtInformation": {
+                        "assigned_to_str": "Lamont C. Berecz"
+                    },
+                },
+            ),
+        ],
+        "juriscraper.opinions.united_states.state.idahoctapp_u": [
+            (
+                "\n           ACADEMY ASSOCIATION, INC.,          )\na North Carolina corporation; and   )\nSPOKANE TEACHERS CREDIT             )\nUNION, a Washington credit union,   )\n                                    )\n       Defendants.                  )\n____________________________________)\n\n\n      Appeal from the District Court of the First Judicial District of the State of Idaho,\n      Bonner County. Lamont C. Berecz, District Judge.\n\n      The district court’s judgment is affirmed.\n\n      Condon Tobin Sladek Thornton Nerenberg, PLLC, Dallas, TX and Parsons Behle &\n      Latimer, Boise, attorneys for Appellant, Global Growth Holdings, Inc. Jared T.S.\n      Pace submitted argument on the briefs.\n\n      Givens Pursley, LLP, Boise, attorneys for Respondent, Erie Properties, LLC.\n      Morgan D. Goodin submitted argument on the briefs.\n",
+                {
+                    "Docket": {
+                        "appeal_from_str": "District Court of the First Judicial District of the State of Idaho, Bonner County",
+                    },
+                    "OriginatingCourtInformation": {
+                        "assigned_to_str": "Lamont C. Berecz"
+                    },
+                },
+            ),
+        ],
         "juriscraper.opinions.united_states.state.iowa": [
             (
                 "                          In the Iowa Supreme Court\n\n                                     No. 23–1122\n\n              Submitted September 11, 2024—Filed October 11, 2024\n                             Amended April 17, 2025\n\n                                    State of Iowa,\n\n                                       Appellee,\n\n                                          vs.\n\n                           Willard Noble Chaiden Miller,\n\n                                      Appellant.\n\n\n         Appeal from the Iowa District Court for Jefferson County, Shawn Showers,\n\njudge.\n\n         A defendant convicted of first-degree murder as a juvenile appeals his\n\nsentence of life imprisonment with the possibility of parole after serving\n\nthirty-five years. Affirmed.\n\n         Christensen, C.J., delivered the opinion of the court, in which all justices\n\njoined.",